Extract walk-in team card data and stat surface hooks
