+++ apps/web/src/features/walk-in/use-walk-in-team-card-page-data.ts
@@ -1,0 +1,190 @@
+import { useMemo } from "react";
+import { useAnalyticsQuery } from "@/features/analytics/queries/useAnalyticsQuery";
+import {
+	type TeamPageMemberRow,
+	useTeamPageData,
+} from "@/features/team/use-team-page-data";
+import type { WalkInOnboardingMetrics } from "@/features/walk-in/walk-in-onboarding-types";
+import {
+	buildResolvedTeamCardRow,
+	buildWalkInOnboardingMetrics,
+	buildWalkInStatItems,
+} from "@/features/walk-in/walk-in-team-card-models";
+import { useWalkInCardData } from "@/features/walk-in/use-walk-in-card-data";
+import type { WalkInTeamMemberCardStatItem } from "@/features/walk-in/WalkInTeamMemberCard";
+import { MAX_ANALYTICS_DAYS } from "@/lib/analytics-date-range";
+import { authClient } from "@/lib/auth-client";
+import { orpc } from "@/lib/orpc";
+
+interface UseWalkInTeamCardPageDataResult {
+	completionUserId: string | null;
+	onboardingMetrics: WalkInOnboardingMetrics;
+	statItems: readonly WalkInTeamMemberCardStatItem[];
+	visibleTeamCardRow: TeamPageMemberRow;
+}
+
+export function useWalkInTeamCardPageData(): UseWalkInTeamCardPageDataResult {
+	const { accountLabel, handover, session, wrappedData } = useWalkInCardData();
+	const { teamMemberRows } = useTeamPageData();
+	const sessionUserId = getSessionUserId(session);
+	const sessionUserName = getSessionUserName(session);
+	const sessionUserEmail = getSessionUserEmail(session);
+	const debugProfileImageSrc = handover.preview.profile.avatarSrc;
+	const { data: activeMember } = authClient.useActiveMember();
+	const activeMemberUserId = getActiveMemberUserId(activeMember);
+	const resolvedUserId = sessionUserId ?? activeMemberUserId;
+	const completionUserId = resolvedUserId ?? null;
+	const developerDetailsQuery = useAnalyticsQuery({
+		...orpc.analytics.developers.details.queryOptions({
+			input: {
+				userId: resolvedUserId ?? "",
+				days: MAX_ANALYTICS_DAYS,
+			},
+		}),
+		enabled: Boolean(resolvedUserId),
+	});
+	const developerFeaturesQuery = useAnalyticsQuery({
+		...orpc.analytics.developers.features.queryOptions({
+			input: {
+				userId: resolvedUserId ?? "",
+				days: MAX_ANALYTICS_DAYS,
+			},
+		}),
+		enabled: Boolean(resolvedUserId),
+	});
+	const developerProjectsQuery = useAnalyticsQuery({
+		...orpc.analytics.developers.projects.queryOptions({
+			input: {
+				userId: resolvedUserId ?? "",
+				days: MAX_ANALYTICS_DAYS,
+			},
+		}),
+		enabled: Boolean(resolvedUserId),
+	});
+	const developerSessionsQuery = useAnalyticsQuery({
+		...orpc.analytics.developers.sessions.queryOptions({
+			input: {
+				userId: resolvedUserId ?? "",
+				days: MAX_ANALYTICS_DAYS,
+				outcome: "all",
+				limit: 1000,
+				offset: 0,
+				sortBy: "date",
+				sortOrder: "desc",
+			},
+		}),
+		enabled: Boolean(resolvedUserId),
+	});
+	const commitBreakdownQuery = useAnalyticsQuery({
+		...orpc.analytics.sessions.dimensionAnalysis.queryOptions({
+			input: {
+				days: MAX_ANALYTICS_DAYS,
+				dimension: "has_commit",
+				limit: 4,
+				metric: "session_count",
+				userId: resolvedUserId ?? undefined,
+			},
+		}),
+		enabled: Boolean(resolvedUserId),
+	});
+	const visibleTeamCardRow = useMemo(
+		() =>
+			buildResolvedTeamCardRow({
+				accountLabel,
+				debugProfileImageSrc,
+				developerDetails: developerDetailsQuery.data,
+				sessionUserEmail,
+				sessionUserId: resolvedUserId,
+				sessionUserName,
+				teamMemberRows,
+			}),
+		[
+			accountLabel,
+			debugProfileImageSrc,
+			developerDetailsQuery.data,
+			sessionUserEmail,
+			resolvedUserId,
+			sessionUserName,
+			teamMemberRows,
+		],
+	);
+	const onboardingMetrics = useMemo(
+		() =>
+			buildWalkInOnboardingMetrics({
+				commitBreakdown: commitBreakdownQuery.data,
+				developerDetails: developerDetailsQuery.data,
+				developerFeatures: developerFeaturesQuery.data,
+				developerProjects: developerProjectsQuery.data,
+				developerSessions: developerSessionsQuery.data,
+				wrappedMetrics: wrappedData?.metrics,
+			}),
+		[
+			commitBreakdownQuery.data,
+			developerDetailsQuery.data,
+			developerFeaturesQuery.data,
+			developerProjectsQuery.data,
+			developerSessionsQuery.data,
+			wrappedData?.metrics,
+		],
+	);
+	const statItems = useMemo(
+		() =>
+			buildWalkInStatItems(
+				visibleTeamCardRow,
+				developerDetailsQuery.data?.distinct_projects ?? 0,
+				wrappedData?.metrics.source_split ?? [],
+			),
+		[
+			developerDetailsQuery.data?.distinct_projects,
+			visibleTeamCardRow,
+			wrappedData?.metrics.source_split,
+		],
+	);
+
+	return {
+		completionUserId,
+		onboardingMetrics,
+		statItems,
+		visibleTeamCardRow,
+	};
+}
+
+function getSessionUserId(
+	session: ReturnType<typeof useWalkInCardData>["session"],
+) {
+	return session?.user &&
+		"id" in session.user &&
+		typeof session.user.id === "string"
+		? session.user.id
+		: undefined;
+}
+
+function getActiveMemberUserId(
+	activeMember: ReturnType<typeof authClient.useActiveMember>["data"],
+) {
+	return activeMember &&
+		"userId" in activeMember &&
+		typeof activeMember.userId === "string"
+		? activeMember.userId
+		: undefined;
+}
+
+function getSessionUserName(
+	session: ReturnType<typeof useWalkInCardData>["session"],
+) {
+	return session?.user &&
+		"name" in session.user &&
+		typeof session.user.name === "string"
+		? session.user.name
+		: undefined;
+}
+
+function getSessionUserEmail(
+	session: ReturnType<typeof useWalkInCardData>["session"],
+) {
+	return session?.user &&
+		"email" in session.user &&
+		typeof session.user.email === "string"
+		? session.user.email
+		: undefined;
+}
